Yesss

+++ struct.h
@@ -8,7 +8,7 @@ typedef struct list_t {
 } list_t;
 
 typedef struct mc_node_t {
-  int32_t wins, sims;
+  int64_t wins, sims;
   double value;
   int8_t turn, child_nr, max_children; // -1 for 0 and 1 for X; move_nr = nr of children
   list_t *child_list;
